refactor(jina): playlist 컴포넌트 widgets으로 이동
