+++ src/widgets/ui/PlayList/index.tsx
@@ -2,6 +2,7 @@ import { PlayListMusicInfo } from './PlayListMusicInfo';
 import { PlayListMusicList } from './PlayListMusicList';
 import { Layout, LeftContainer, RightContainer } from './styled';
 
+// TODO config, model, features로 분리
 export const PlayListPage = () => {
   return (
     <Layout>
